Add simple styling to existing elements
Add the font import for the two given fonts in the instructions. Abscent any
other instructions, Robot will be the page default and Oswald for actionable,
and other differentiated elements.

The given color rgb(226,28,35) is for the actionable items. Complementary colors
will be found for other things. Non-true white will replace the normal black and
white stuff. The company name in normal text will be reversed, as the real site
seems to have as the normal.

Also, need to be aware of the need to print the page, it is a press release
after all.

The elements labeled as <section> didn't seem big enough, nor did the entire
page, to justify their semanitics. Switched to <div>s instead.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>