Add images to the page
As a press release, the copy is rather sparse and dressing it up with an image
just seemed like a good move. Decided to use two different images for print and
screen, with the aspect ratio being larger for the print version, claiming less
of the printed page.

The print image is grayscaled to make it a better option for doing printing and
copies produced offline.

Moved the styles around, placing the @media queries closer to the document, and
thus higher in priority for rendering choices.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>